+++ main/simulator.h
@@ -38,7 +38,7 @@ public:
       : tracer_(tracer), model_(model), source_(source),
         offsetter_(offsetter){};
 
-  std::vector<float> run(float frequency, int numCollectors) {
+  std::vector<double> run(double frequency, int numCollectors) {
     std::vector<std::unique_ptr<objects::EnergyCollector>> collectors =
         buildCollectors(*model_, numCollectors);
   }
